requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.2.0
coloredlogs==15.0.1
Django==4.2.2
django-cors-headers==4.2.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
flatbuffers==23.5.26
graphqlclient==0.2.4
humanfriendly==10.0
idna==3.4
lazy_loader==0.3
llvmlite==0.40.1
mpmath==1.3.0
networkx==3.1
numba==0.57.1
numpy==1.24.4
onnxruntime==1.15.1
packaging==23.1
platformdirs==3.8.1
pooch==1.7.0
protobuf==4.23.4
pyactiveresource==2.2.2
PyMatting==1.1.8
pytz==2020.1
PyWavelets==1.4.1
PyYAML==5.3.1
requests==2.31.0
scipy==1.11.1
ShopifyAPI==12.3.0
six==1.15.0
sqlparse==0.3.1
sympy==1.12
tifffile==2023.7.10
tqdm==4.65.0
typing_extensions==4.6.3
urllib3==2.0.3
